+++ CHANGELOG.md
@@ -41,6 +41,8 @@
 * Validation of standard encryption dictionary to auto-correct invalid /U and
   /O fields in case they are padded with zeros
 * [HexaPDF::Document#wrap] handling of sub-type mapping in case of missing type
+* [HexaPDF::Type::AcroForm::AppearanceGenerator] to also take a text field
+  widget's width into account when auto-sizing
 
 
 ## 0.32.2 - 2023-05-06